Set the foundation for the XML viewer
In order to show the native XML as part of the web document, it
will be necessary to copy the elements and attributes into a <span>
which can then be hidden or shown at will. The stylesheet
xmlViewer.xsl performs this step, but does not attempt to make any
assumptions about what HTML elements should result from the
original XML document. A separate XSL stylesheet should make those
conversions, and import xmlViewer.xsl to handle the XML code
itself.