On full XPath completion, remove styling on non-matches
Also, nodes are tested before being added to pathStep.axisPopulace.
If a node's element is already present in the array, there is no
need to add it again.

If no nodes were added, pathStep.axisPopulace is reset to an empty
array. As a result, pathStep.isComplete() can now return true on
the descendant axis.